backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
numpy==2.3.4
gunicorn==23.0.0
whitenoise==6.8.2
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.3.0